backend/requirements.txt
flask
prometheus_client
flask-cors
flask_sqlalchemy
psycopg2-binary
requests
python-dotenv
